refactor(api): move repositories  into prescription
