+++ api/scripts/prod/target-profile-migrations/migrate-stages-to-level.js
@@ -10,10 +10,10 @@ import { utils as xlsxUtils, writeXLSX } from 'xlsx';
 
 import { disconnect } from '../../../db/knex-database-connection.js';
 import * as stageCollectionRepository from '../../../src/evaluation/infrastructure/repositories/stage-collection-repository.js';
+import * as targetProfileAdministrationRepository from '../../../src/prescription/target-profile/infrastructure/repositories/target-profile-administration-repository.js';
 import { learningContentCache as cache } from '../../../src/shared/infrastructure/caches/learning-content-cache.js';
 import * as organizationRepository from '../../../src/shared/infrastructure/repositories/organization-repository.js';
 import * as skillRepository from '../../../src/shared/infrastructure/repositories/skill-repository.js';
-import * as targetProfileForAdminRepository from '../../../src/shared/infrastructure/repositories/target-profile-for-admin-repository.js';
 import { logger } from '../../../src/shared/infrastructure/utils/logger.js';
 
 const modulePath = fileURLToPath(import.meta.url);
@@ -35,7 +35,7 @@ async function _computeMigrations(inputFile) {
 
   return Promise.all(
     targetProfileIds.map(async (targetProfileId) => {
-      const targetProfile = await targetProfileForAdminRepository.get({ id: targetProfileId });
+      const targetProfile = await targetProfileAdministrationRepository.get({ id: targetProfileId });
 
       const levels = await _computeLevels(targetProfile.cappedTubes);
