+++ api/scripts/prod/target-profile-migrations/revert-stages-to-threshold.js
@@ -10,8 +10,8 @@ import { readFile, set_fs, utils as xlsxUtils } from 'xlsx';
 
 import { disconnect } from '../../../db/knex-database-connection.js';
 import * as stageCollectionRepository from '../../../src/evaluation/infrastructure/repositories/stage-collection-repository.js';
+import * as targetProfileAdministrationRepository from '../../../src/prescription/target-profile/infrastructure/repositories/target-profile-administration-repository.js';
 import { learningContentCache as cache } from '../../../src/shared/infrastructure/caches/learning-content-cache.js';
-import * as targetProfileForAdminRepository from '../../../src/shared/infrastructure/repositories/target-profile-for-admin-repository.js';
 import { logger } from '../../../src/shared/infrastructure/utils/logger.js';
 
 set_fs(fs);
@@ -43,7 +43,9 @@ async function _computeReverts(inputFile, sample) {
     logger.info({ targetProfileIds }, `Using sample target profiles`);
   }
 
-  const targetProfiles = await Promise.all(targetProfileIds.map((id) => targetProfileForAdminRepository.get({ id })));
+  const targetProfiles = await Promise.all(
+    targetProfileIds.map((id) => targetProfileAdministrationRepository.get({ id })),
+  );
 
   return targetProfiles
     .map((targetProfile) => {
