+++ api/tests/unit/domain/usecases/get-target-profile-for-admin_test.js
@@ -4,16 +4,16 @@ import { expect, sinon } from '../../../test-helper.js';
 describe('Unit | UseCase | get-target-profile-for-admin', function () {
   it('should get target profile for admin', async function () {
     // given
-    const targetProfileForAdminRepository = {
+    const targetProfileAdminRepository = {
       get: sinon.stub(),
     };
     const targetProfile = Symbol('targetProfile');
-    targetProfileForAdminRepository.get.withArgs({ id: 123 }).resolves(targetProfile);
+    targetProfileAdminRepository.get.withArgs({ id: 123 }).resolves(targetProfile);
 
     // when
     const result = await usecases.getTargetProfileForAdmin({
       targetProfileId: 123,
-      targetProfileForAdminRepository,
+      targetProfileAdminRepository,
     });
 
     // then
